list joined events, de-reg events done
