+++ app/src/main/java/com/example/stu_share/EventJoinedDetail.java
@@ -16,6 +16,20 @@ import android.widget.Toast;
 
 import com.google.android.material.bottomnavigation.BottomNavigationView;
 
+import org.json.JSONException;
+import org.json.JSONObject;
+
+import java.io.BufferedReader;
+import java.io.BufferedWriter;
+import java.io.DataInputStream;
+import java.io.DataOutputStream;
+import java.io.IOException;
+import java.io.InputStreamReader;
+import java.io.OutputStreamWriter;
+import java.net.HttpURLConnection;
+import java.net.ProtocolException;
+import java.net.URL;
+
 import butterknife.BindView;
 import butterknife.ButterKnife;
 
@@ -27,6 +41,7 @@ public class EventJoinedDetail extends AppCompatActivity {
     ImageView buttonImg;
     @BindView(R.id.toolbar)
     public Toolbar toolBar;
+    private String url="https://w0044421.gblearn.com/stu_share/EventDeReg.php";
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
@@ -57,7 +72,6 @@ public class EventJoinedDetail extends AppCompatActivity {
                         break;
                     case R.id.action_message:
                         Intent intent = new Intent(getBaseContext(), MessageList.class);
-//              intent.putExtra("args", userReg);
                         intent.putExtra("user",user);
                         startActivity(intent);
                         break;
@@ -96,13 +110,61 @@ public class EventJoinedDetail extends AppCompatActivity {
 
                 Toast.makeText(getBaseContext(), "Deregister successfully!",
                         Toast.LENGTH_LONG).show();
-                Intent intent=new Intent(getBaseContext(),EventList.class);
+                dereg(url,event,user);
+                Intent intent=new Intent(getBaseContext(),EventListJoined.class);
                 intent.putExtra("user",user);
                 intent.putExtra("event",event);
                 startActivity(intent);
             }
         });
     }
+
+    private void dereg(final String url, final EventCoordinator.Event event, final User user) {
+        Thread thread = new Thread(new Runnable() {
+            @Override
+            public void run() {
+                try {
+                    URL url1 = new URL(url);
+                    HttpURLConnection conn = (HttpURLConnection) url1.openConnection();
+                    conn.setRequestMethod("POST");
+                    conn.setRequestProperty("Content-Type", "application/json;charset=UTF-8");
+                    conn.setRequestProperty("Accept", "application/json");
+                    conn.setDoOutput(true);
+                    conn.setDoInput(true);
+                    JSONObject jsonParam = new JSONObject();
+                    jsonParam.put("userId", user.id);
+                    jsonParam.put("eventId", event.id);
+                    Log.i("JSON", jsonParam.toString());
+                    DataOutputStream os = new DataOutputStream(conn.getOutputStream());
+                    BufferedWriter writer = new BufferedWriter(
+                            new OutputStreamWriter(os, "UTF-8"));
+                    os.writeBytes(jsonParam.toString());
+                    os.flush();
+                    os.close();
+                    conn.connect();
+                    Log.i("STATUS", String.valueOf(conn.getResponseCode()));
+                    Log.i("MSG", conn.getResponseMessage());
+                    BufferedReader in = new BufferedReader(new InputStreamReader(conn.getInputStream()));
+                    DataInputStream is = new DataInputStream(conn.getInputStream());
+                    final StringBuilder total = new StringBuilder();
+                    String line;
+                    while ((line = in.readLine()) != null) {
+                        total.append(line).append('\n');
+                    }
+                    Log.d("TAG", "Server Response is: " + total.toString() + ": ");
+
+                } catch (ProtocolException e) {
+                    e.printStackTrace();
+                } catch (IOException e) {
+                    e.printStackTrace();
+                } catch (JSONException e) {
+                    e.printStackTrace();
+                }
+            }
+        });
+        thread.start();
+    }
+
     public void OpenMenuActivity() {
         Intent intent = new Intent(this, EventList.class);
         intent.putExtra("user",user);
