--- app/src/main/res/layout/activity_event_reg_detail.xml
@@ -1,189 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:background="@drawable/bkgimg"
-    tools:context=".EventRegDetail">
-
-    <androidx.appcompat.widget.Toolbar
-        android:id="@+id/toolbar"
-        android:layout_width="match_parent"
-        android:layout_height="?attr/actionBarSize"
-        android:background="#2F3C81"
-        app:layout_collapseMode="pin"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-
-        app:layout_constraintTop_toTopOf="parent"
-        app:layout_scrollFlags="scroll|enterAlways"
-        app:popupTheme="@style/ThemeOverlay.AppCompat.Light"
-        app:theme="@style/ThemeOverlay.AppCompat.Dark.ActionBar">
-
-        <androidx.appcompat.widget.AppCompatImageView
-            android:id="@+id/buttonImg"
-            android:layout_width="200px"
-            android:layout_height="150px"
-            android:layout_gravity="right"
-            android:layout_marginRight="@dimen/activity_horizontal_margin"
-
-
-
-            android:src="@drawable/ic_create" />
-
-    </androidx.appcompat.widget.Toolbar>
-
-    <TextView
-        android:id="@+id/txtEndTime123"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginTop="16dp"
-        android:text="text end time"
-        android:textColor="#ECE9E5"
-        android:textSize="16dp"
-        app:layout_constraintStart_toEndOf="@+id/textView124"
-        app:layout_constraintTop_toBottomOf="@+id/txtStTime123" />
-
-    <TextView
-        android:id="@+id/txtStDate123"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginTop="8dp"
-        android:text="start date"
-        android:textColor="#ECE9E5"
-        android:textSize="16dp"
-        app:layout_constraintStart_toEndOf="@+id/textView128"
-        app:layout_constraintTop_toTopOf="@+id/textView128" />
-
-    <TextView
-        android:id="@+id/txtStTime123"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginTop="8dp"
-        android:text="start time"
-        android:textColor="#ECE9E5"
-        android:textSize="16dp"
-        app:layout_constraintStart_toEndOf="@+id/textView123"
-        app:layout_constraintTop_toTopOf="@+id/textView128" />
-
-    <TextView
-        android:id="@+id/txtEndDate123"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginTop="12dp"
-        android:text="end date"
-        android:textColor="#ECE9E5"
-        android:textSize="16dp"
-        app:layout_constraintEnd_toStartOf="@+id/textView13"
-        app:layout_constraintHorizontal_bias="0.0"
-        app:layout_constraintStart_toEndOf="@+id/textView129"
-        app:layout_constraintTop_toBottomOf="@+id/textView128" />
-
-    <TextView
-        android:id="@+id/textView120"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginStart="20dp"
-        android:layout_marginTop="39dp"
-        android:text="Event"
-        android:textAppearance="@style/TextAppearance.AppCompat.Large"
-        android:textColor="#ECE9E5"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/btnHome" />
-
-    <TextView
-        android:id="@+id/txtEventTitle123"
-        android:layout_width="wrap_content"
-        android:layout_height="26dp"
-        android:layout_marginStart="20dp"
-        android:layout_marginLeft="20dp"
-        android:textSize="16dp"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/textView120" />
-
-    <TextView
-        android:id="@+id/textView127"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginStart="20dp"
-        android:layout_marginTop="60dp"
-        android:text="Detail"
-        android:textAppearance="@style/TextAppearance.AppCompat.Large"
-        android:textColor="#ECE9E5"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/txtEventTitle123" />
-
-    <TextView
-        android:id="@+id/textView123"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginStart="174dp"
-        android:text="Time"
-        android:textAppearance="@style/TextAppearance.AppCompat.Large"
-        android:textColor="#ECE9E5"
-        app:layout_constraintStart_toEndOf="@+id/textView128"
-        app:layout_constraintTop_toTopOf="@+id/textView128" />
-
-    <TextView
-        android:id="@+id/textView124"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginTop="6dp"
-        android:text="Time"
-        android:textAppearance="@style/TextAppearance.AppCompat.Large"
-        app:layout_constraintStart_toStartOf="@+id/textView123"
-        app:layout_constraintTop_toBottomOf="@+id/textView123"
-        android:textColor="#ECE9E5" />
-
-    <TextView
-        android:id="@+id/txtEvtDetail123"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginStart="20dp"
-        android:layout_marginLeft="20dp"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/textView127" />
-
-    <TextView
-        android:id="@+id/textView128"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginStart="20dp"
-        android:layout_marginTop="20dp"
-        android:text="Start"
-        android:textAppearance="@style/TextAppearance.AppCompat.Large"
-        android:textColor="#ECE9E5"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/textView120" />
-
-    <TextView
-        android:id="@+id/textView129"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginStart="20dp"
-        android:layout_marginTop="6dp"
-        android:text="End"
-        android:textAppearance="@style/TextAppearance.AppCompat.Large"
-        android:textColor="#ECE9E5"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/textView128" />
-
-    <Button
-        android:id="@+id/btnReg123"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginTop="50dp"
-        android:layout_marginEnd="20dp"
-        android:layout_marginRight="20dp"
-        android:text="DeRegister"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/txtEvtDetail123" />
-
-    <include
-        layout="@layout/bottom_navigation"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_gravity="bottom"
-        app:layout_constraintBottom_toBottomOf="parent" />
-</androidx.constraintlayout.widget.ConstraintLayout>
